Moved Game to its own file to avoid clutter
